requirements.txt
streamlit
opencv-python-headless
numpy
pandas
paho-mqtt
Pillow
